# Speed up processing of CSV files