@leohemsted 2016-07-07
not sure if it'll make a big difference now that it's not being called as much, but not splitting and rejoining has a factor of 10 improvement

```
(api) gds0260:dev leohemsted$ python -m timeit "''.join(c.lower() for c in 'PhOne_number -' if c not in ' _-')"
100000 loops, best of 3: 3.58 usec per loop
(api) gds0260:dev leohemsted$ python -m timeit "'PhOne_number -'.lower().replace(' _-', '')"
1000000 loops, best of 3: 0.364 usec per loop
```